test(reader): split reader source guards
Why:
- Continue the test architecture migration by moving package-owned reader source guards out of the app-host XCTest bundle.
- Keep Android/Vue quick-selector and active-window parity protected without paying app launch cost for source-contract checks.

What Changed:
- Added ReaderSourceGuardTests under the BibleUI package test target for toolbar quick-selector and active-window source contracts.
- Centralized Swift function source extraction in BibleUITestSourceLocator and reused it from the module-picker source guard.
- Removed the migrated source guards and duplicate helper code from AndBibleTests+AppAndReader.
- Updated docs/test-architecture-migration-plan.md to reflect the remaining app-reader migration work.

Validation:
- git diff --cached --check
- python3 scripts/check_repo_standards.py docblocks --all-files
- python3 scripts/check_repo_standards.py source-guards
- GitNexus detect_changes scope=staged
- xcodebuild test -project AndBible.xcodeproj -scheme BibleUITests -destination 'platform=iOS Simulator,name=iPhone 17' -derivedDataPath /tmp/andbible-reader-source-guards-bibleui-dd CODE_SIGNING_ALLOWED=NO -only-testing:BibleUITests/ReaderSourceGuardTests -only-testing:BibleUITests/BibleReaderModulePickerTests/testBibleReaderModulePickerRoutesMapsThroughDocumentSwitch
- xcodebuild test -quiet -project AndBible.xcodeproj -scheme AndBible -destination 'platform=iOS Simulator,name=iPhone 17' -derivedDataPath /tmp/andbible-reader-source-guards-app-dd CODE_SIGNING_ALLOWED=NO -only-testing:AndBibleTests/AndBibleTests/testBibleReaderSpeakMiniPlayerBuildsWithSpeakService -only-testing:AndBibleTests/AndBibleTests/testApplicationDelegateSceneConfigurationUsesWindowSceneDelegate
- Adversarial sub-agent review found no staged-diff issues.

Impact:
- No production behavior changes.
- Reduces app-host test scope while preserving parity guards in the owning package target.
- Leaves inherited source-string fragility for a future contract-test hardening effort.
