+++ Sources/BibleUI/Tests/BibleUITests/BibleUITestSourceLocator.swift
@@ -53,4 +53,54 @@ enum BibleUITestSourceLocator {
         let sourceURL = root.appendingPathComponent(relativePath)
         return try String(contentsOf: sourceURL, encoding: .utf8)
     }
+
+    /**
+     Extracts one Swift function body from source text for focused source-contract assertions.
+
+     Source guards should stay tied to the function boundary they protect instead of searching for
+     arbitrary file-wide fragments. The scanner balances braces after the named function
+     declaration and returns only that function's source body.
+
+     - Parameter functionName: Name of the Swift function declaration to extract.
+     - Parameter source: Source file contents containing the function declaration.
+     - Returns: Source text from the function declaration through its closing brace.
+     - Throws: An `NSError` when the declaration is missing or the function body is unbalanced.
+     */
+    static func extractFunction(named functionName: String, from source: String) throws -> String {
+        guard let functionRange = source.range(of: "func \(functionName)") else {
+            throw NSError(
+                domain: "BibleUITestSourceLocator",
+                code: 2,
+                userInfo: [NSLocalizedDescriptionKey: "Function \(functionName) not found"]
+            )
+        }
+        guard let openingBrace = source[functionRange.lowerBound...].firstIndex(of: "{") else {
+            throw NSError(
+                domain: "BibleUITestSourceLocator",
+                code: 3,
+                userInfo: [NSLocalizedDescriptionKey: "Function \(functionName) has no body"]
+            )
+        }
+
+        var depth = 0
+        var current = openingBrace
+        while current < source.endIndex {
+            let character = source[current]
+            if character == "{" {
+                depth += 1
+            } else if character == "}" {
+                depth -= 1
+                if depth == 0 {
+                    return String(source[functionRange.lowerBound...current])
+                }
+            }
+            current = source.index(after: current)
+        }
+
+        throw NSError(
+            domain: "BibleUITestSourceLocator",
+            code: 4,
+            userInfo: [NSLocalizedDescriptionKey: "Function \(functionName) body was not balanced"]
+        )
+    }
 }
